Remove import guard from PEtab benchmark
