+++ tests/benchmark-models/test_petab_benchmark.py
@@ -6,39 +6,38 @@ for a subset of the benchmark problems.
 """
 
 import copy
-from functools import partial
 from pathlib import Path
 
+import contextlib
+import logging
 import os
-import fiddy
-import amici
+from collections import defaultdict
+from dataclasses import dataclass, field
+
 import numpy as np
 import pandas as pd
 import petab.v1 as petab
 import pytest
-from amici.petab.petab_import import import_petab_problem
-import benchmark_models_petab
-from collections import defaultdict
-from dataclasses import dataclass, field
-from amici import SensitivityMethod
-from petab.v1.lint import measurement_table_has_timepoint_specific_mappings
-from fiddy import MethodId, get_derivative
-from fiddy.derivative_check import NumpyIsCloseDerivativeCheck
-from fiddy.extensions.amici import simulate_petab_to_cached_functions
-from fiddy.success import Consistency
-import contextlib
-import logging
 import yaml
+from petab.v1.lint import measurement_table_has_timepoint_specific_mappings
+from petab.v1.visualize import plot_problem
+
+import amici
+from amici import SensitivityMethod
 from amici.logging import get_logger
+from amici.petab.petab_import import import_petab_problem
 from amici.petab.simulations import (
     LLH,
-    SLLH,
     RDATAS,
     rdatas_to_measurement_df,
     simulate_petab,
 )
-
-from petab.v1.visualize import plot_problem
+import benchmark_models_petab
+import fiddy
+from fiddy import MethodId, get_derivative
+from fiddy.derivative_check import NumpyIsCloseDerivativeCheck
+from fiddy.extensions.amici import simulate_petab_to_cached_functions
+from fiddy.success import Consistency
 
 
 # Enable various debug output
@@ -259,113 +258,6 @@ def benchmark_problem(request):
         model_output_dir=benchmark_outdir / problem_id,
     )
     return problem_id, flat_petab_problem, petab_problem, amici_model
-
-
-@pytest.mark.filterwarnings(
-    "ignore:The following problem parameters were not used *",
-    "ignore: The environment variable *",
-    "ignore:Adjoint sensitivity analysis for models with discontinuous ",
-)
-def test_jax_llh(benchmark_problem):
-    import jax
-    import equinox as eqx
-    import jax.numpy as jnp
-    from amici.jax.petab import run_simulations, JAXProblem
-
-    jax.config.update("jax_enable_x64", True)
-    from beartype import beartype
-
-    problem_id, flat_petab_problem, petab_problem, amici_model = (
-        benchmark_problem
-    )
-
-    amici_solver = amici_model.getSolver()
-    cur_settings = settings[problem_id]
-    amici_solver.setAbsoluteTolerance(1e-8)
-    amici_solver.setRelativeTolerance(1e-8)
-    amici_solver.setMaxSteps(10_000)
-
-    simulate_amici = partial(
-        simulate_petab,
-        petab_problem=flat_petab_problem,
-        amici_model=amici_model,
-        solver=amici_solver,
-        scaled_parameters=True,
-        scaled_gradients=True,
-        log_level=logging.DEBUG,
-    )
-
-    np.random.seed(cur_settings.rng_seed)
-
-    problem_parameters = None
-    if problem_id in problems_for_gradient_check:
-        point = flat_petab_problem.x_nominal_free_scaled
-        for _ in range(20):
-            amici_solver.setSensitivityMethod(amici.SensitivityMethod.adjoint)
-            amici_solver.setSensitivityOrder(amici.SensitivityOrder.first)
-            amici_model.setSteadyStateSensitivityMode(
-                cur_settings.ss_sensitivity_mode
-            )
-            point_noise = (
-                np.random.randn(len(point)) * cur_settings.noise_level
-            )
-            point += point_noise  # avoid small gradients at nominal value
-
-            problem_parameters = dict(
-                zip(flat_petab_problem.x_free_ids, point)
-            )
-
-            r_amici = simulate_amici(
-                problem_parameters=problem_parameters,
-            )
-            if np.isfinite(r_amici[LLH]):
-                break
-        else:
-            raise RuntimeError("Could not compute expected derivative.")
-    else:
-        r_amici = simulate_amici()
-    llh_amici = r_amici[LLH]
-
-    jax_model = import_petab_problem(
-        petab_problem,
-        model_output_dir=benchmark_outdir / (problem_id + "_jax"),
-        jax=True,
-    )
-    jax_problem = JAXProblem(jax_model, petab_problem)
-    if problem_parameters:
-        jax_problem = eqx.tree_at(
-            lambda x: x.parameters,
-            jax_problem,
-            jnp.array(
-                [problem_parameters[pid] for pid in jax_problem.parameter_ids]
-            ),
-        )
-
-    if problem_id in problems_for_gradient_check:
-        beartype(run_simulations)(jax_problem)
-        (llh_jax, _), sllh_jax = eqx.filter_value_and_grad(
-            run_simulations, has_aux=True
-        )(jax_problem)
-    else:
-        llh_jax, _ = beartype(run_simulations)(jax_problem)
-
-    np.testing.assert_allclose(
-        llh_jax,
-        llh_amici,
-        rtol=1e-3,
-        atol=1e-3,
-        err_msg=f"LLH mismatch for {problem_id}",
-    )
-
-    if problem_id in problems_for_gradient_check:
-        sllh_amici = r_amici[SLLH]
-        np.testing.assert_allclose(
-            sllh_jax.parameters,
-            np.array([sllh_amici[pid] for pid in jax_problem.parameter_ids]),
-            rtol=1e-2,
-            atol=1e-2,
-            err_msg=f"SLLH mismatch for {problem_id}, {dict(zip(jax_problem.parameter_ids, sllh_jax.parameters))}",
-        )
 
 
 @pytest.mark.filterwarnings(
